chore: relocate `SlimefunGuideUnlockProvider`
