+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/guide/unlockprovider/CurrencyUnlockProvider.java
@@ -5,7 +5,7 @@ import javax.annotation.Nonnull;
 import org.bukkit.entity.Player;
 
 import io.github.thebusybiscuit.slimefun4.api.researches.Research;
-import io.github.thebusybiscuit.slimefun4.core.guide.SlimefunGuideUnlockProvider;
+import io.github.thebusybiscuit.slimefun4.api.guide.SlimefunGuideUnlockProvider;
 import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
 import io.github.thebusybiscuit.slimefun4.integrations.VaultIntegration;
 
@@ -14,7 +14,7 @@ import io.github.thebusybiscuit.slimefun4.integrations.VaultIntegration;
  *
  * @see SlimefunGuideUnlockProvider
  */
-public class CurrencyUnlockProvider implements SlimefunGuideUnlockProvider {
+public final class CurrencyUnlockProvider implements SlimefunGuideUnlockProvider {
     @Override
     public boolean canUnlock(@Nonnull Research research, @Nonnull Player p) {
         if (!VaultIntegration.isAvailable()) {
